+++ mle_logging/__init__.py
@@ -1,18 +1,16 @@
 from ._version import __version__
 from .mle_logger import MLELogger
-from .meta_log import MetaLog
-from .load import load_log, load_meta_log, load_model
-from .merge import merge_config_logs, merge_seed_logs
+from .load import load_log, load_model
 from .utils import load_config
+from .merge import merge_config_logs, merge_seed_logs
+
 
 __all__ = [
     "__version__",
     "MLELogger",
-    "MetaLog",
     "load_log",
-    "load_meta_log",
     "load_model",
+    "load_config",
     "merge_config_logs",
     "merge_seed_logs",
-    "load_config",
 ]